fix(router): move login & register pages to resolve conflict
